+++ crates/api/src/atom/v2/dim.rs
@@ -1,0 +1,7 @@
+use crate::{v2::dim::Dimension, Atom, Composable, Composition};
+
+impl Dimension<Composition> for Atom {
+    fn dimension(&self) -> Composition {
+        Composable::composition(self)
+    }
+}